Converting to pnpm, upgrading lerna, and setting up nx in preparation for building and releasing the 6.x branch using our current CI setup. Removing empty turf-geojson-rbush directory and getting monorepolint upgraded and the 6.5.0 checks ported.
