+++ .gitignore
@@ -32,3 +32,5 @@ npm-debug.log*
 node_modules
 
 .vscode
+
+.nx/
