+++ .prettierignore
@@ -10,3 +10,14 @@ packages/turf-directional-mean/test/in/bus_route_utm.json
 
 # is actually output
 packages/turf/turf.min.js
+
+# Ignore test fixture json in case intentional line breaks help with coord
+# readability.
+packages/turf-*/test/in/**
+packages/turf-*/test/out/**
+packages/turf-*/test/true/**
+packages/turf-*/test/false/**
+
+pnpm-lock.yaml
+
+/.nx
